moved yaml and XML stuff to flowcontrol
