+++ src/flowcontrol/CMakeLists.txt
@@ -6,15 +6,14 @@ add_library(rs_flowcontrol
     src/RSControledAnalysisEngine.cpp
     src/RSProcessManager.cpp
     src/RSParallelPipelinePlanner.cpp
-    src/RSAggregatedAnalysisEngine.cpp)
+    src/RSAggregatedAnalysisEngine.cpp
+    src/RSXMLParser.cpp
+    src/YamlToXMLConverter.cpp)
 add_dependencies(rs_flowcontrol robosherlock_msgs_generate_messages_cpp)
 target_link_libraries(rs_flowcontrol rs_core rs_io rs_queryanswering
   ${LIBAPR_LIBRARY} ${SWIPL_LIBRARIES} ${UIMA_LIBRARY} ${ICUUC_LIBRARY} ${catkin_LIBRARIES}
   ${YAML_CPP_LIBRARIES})
 
 
-
-rs_add_executable(run
-        src/run.cpp)
-target_link_libraries(run
-        rs_flowcontrol ${catkin_LIBRARIES})
+rs_add_executable(run src/run.cpp)
+target_link_libraries(run rs_flowcontrol ${catkin_LIBRARIES})
