+++ src/utils/CMakeLists.txt
@@ -7,8 +7,6 @@ add_library(rs_utils SHARED
   src/WeightedGraph.cpp
   src/DirectedGraph.cpp
   src/GraphPrimitives.cpp
-  src/YamlToXMLConverter.cpp
-  src/RSXMLParser.cpp
 )
 target_link_libraries(rs_utils rs_core ${YAML_CPP_LIBRARIES} ${catkin_LIBRARIES})
